### 2025-07-18 bf49f61
Reduce words number
