+++ paper.md
@@ -31,13 +31,13 @@ bibliography: paper.bib
 # Summary
 
 `Lenapy` is a Python library designed to facilitate the processing and analysis of geophysical and climate datasets, such as those used in oceanography, geodesy, and Earth observation in general. 
-Built on top of `xarray` and fully compatible with `dask`, it enables scalable and reproducible workflows on multidimensional datasets by leveraging community standards for data formats (NetCDF) and metadata conventions (CF).
+Built on top of `xarray` and fully compatible with `dask`, it enables scalable workflows on multidimensional datasets using community standards for data formats (NetCDF) and metadata conventions (CF).
 
-`Lenapy` provides high-level accessors that extend `xarray.Dataset` and `xarray.DataArray` objects, allowing direct application of specialized methods.
+`Lenapy` provides high-level accessors that extend `xarray.Dataset` and `xarray.DataArray` objects, allowing application of specialized methods.
 
-One of `Lenapy`’s key features is its unified approach for Global Mean Sea Level (GMSL) estimation, enabling consistent computation from steric, manometric, or relative sea level components.
+One of `Lenapy`’s key features is its unified approach for Global Mean Sea Level estimation, enabling consistent computation of its components.
 
-`Lenapy` is intended to support Earth system scientists, oceanographers, and climate service providers with unified and coherent tools through an extensible and transparent interface.
+`Lenapy` is intended to support geophysicists, oceanographers, and climate service providers with coherent tools through an extensible interface.
 
 
 
@@ -51,7 +51,7 @@ The Thermodynamic Equation of Seawater (TEOS-10) framework provides consistent d
 Existing libraries address specific aspects of these requirements:
 
 - `pyshtools` [@Wieczorek_2018] is a comprehensive package for spherical harmonic transforms and spectral analysis, particularly in geophysics. Yet, it operates on standalone arrays and does not natively support xarray, limiting its compatibility with NetCDF-based workflows.
-- `grates` [@Kvas] provides object-oriented tools for spherical harmonics and geodetic computations, but similarly lacks integration with labeled multidimensional data structures like xarray.
+- `grates` [@Kvas] provides object-oriented tools for spherical harmonics, but similarly only operates on standalone arrays.
 - `gsw-xarray` [@Caneill_2024] provides implementations of the TEOS-10, facilitating oceanographic computations. It offers a wrapper around `GSW-Python` for xarray objects, but without Dask support.
 
 While `gsw-xarray` is more complete than our GSW wrapper in `Lenapy`, our library propose complementary geodetic tools for spatial and spherical harmonics operations.
@@ -65,8 +65,8 @@ For example, users can compute area-weighted means via `ds.lngeo.mean()` or extr
 `Lenapy` is designed for Earth scientists, oceanographers, climate researchers, and geodesists who routinely manipulate global or regional gridded datasets and require specific processing workflows.
 `Lenapy` aims to maintain full compatibility with the PyData ecosystem.
 
-Furthermore, `Lenapy` offers a unified approach for calculating Global Mean Sea Level (GMSL) by integrating both steric and manometric components, as well as relative sea-level changes [@Gregory_2019].
-`Lenapy` facilitates this decomposition by providing a unique Python library that compute these components directly from xarray-based datasets, enabling researchers to analyze sea-level changes comprehensively within a consistent framework.
+Furthermore, `Lenapy` offers a unified approach for calculating Global Mean Sea Level by integrating both steric and manometric components, as well as relative sea-level changes [@Gregory_2019].
[…]
