### 2025-08-01 481a223
bug climato gestion Nmin
