### 2025-08-01 5e48c98
bug climato Nmin
