### 2025-08-19 #8602 @hulecom
Hi @kthyng

Is it possible to have an update on the current status of the pre-review please ?
Do we have to take any action on our side (or just have to wait) ?

Thanks
