### 2025-09-02 b494d4f
test: add lnocean.above unittest
