### 2025-09-02 3968cbb
refact: refactor OceanArray.above
