### 2025-09-02 2b1834b
test: wip, add climato() unit test with chunks
