+++ tests/utils/test_climato.py
@@ -15,3 +15,8 @@ def test_climato_coeffs(lenapy_paths):
     assert (
         sorted(clim.coeff_names) == ref_coeffs
     ), "Coeffs_climato outputs do not correspond to the ref."
+
+
+def test_climato(lenapy_paths):
+    ohc_data = xr.open_dataset(lenapy_paths.data / "ohc.nc", chunks=True)
+    clim = ohc_data.gohc.lntime.climato()
