### 2025-09-09 #8602 @kthyng
@hulecom
* Can you release your Python package on PyPI or conda-forge?
