### 2025-09-09 #8602 @hulecom
Hi @kthyng

I realized that the 'joss' branch of our repository was not up to date with the 'main' branch so I just fixed that with a merge.

We deployed Lenapy package on conda-forge ([https://anaconda.org/conda-forge/lenapy](https://anaconda.org/conda-forge/lenapy)) and installation instructions have been updated accordingly on the documentation / Readme.md. The procedure has started for PyPi release (I don't know the current status yet, I am not in charge of it).
