### 2025-09-11 367eedd
enh: add dim chunk check
