### 2025-09-11 a6cb6a9
enh: add dim chunk check
