### 2025-09-25 #8602 @elbeejay
@hulecom I'll be taking a look and contacting potential peer-reviewers over the next few days. I appreciate your patience; clearly we are a bit backlogged at JOSS !
