### 2025-09-28 #8602 @elbeejay
@ThomasMGeo can you review this submission titled, _Lenapy: Enhancing xarray for multidimensional geophysical data analysis_ for JOSS? I think it's up your alley as it relates to programming with xarray and geophysics.
