### 2025-09-28 #8602 @elbeejay
@nicrie can you review this submission titled, _Lenapy: Enhancing xarray for multidimensional geophysical data analysis_ for JOSS? Given your JOSS publication on `xeofs`, you seem like an ideal peer-reviewer for `Leanapy`.
