### 2025-09-28 #8602 @elbeejay
@ehinman do you think you or one of your colleagues would be interested in reviewing this submission titled, _Lenapy: Enhancing xarray for multidimensional geophysical data analysis_ for JOSS? I think use and application of `xarray` to geo-data might be interesting and beneficial for folks in your line of work to look into; and exposure/learning by reviewing could be fulfilling. Please take a look and let me know; reach out if you have any questions about [JOSS](https://joss.theoj.org/) or our [review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
