### 2025-09-29 #8602 @nicrie
Thanks @elbeejay for the ping! But I won’t be able to carve out enough time in the coming months to do a proper review, so I’ll have to pass.
