### 2025-10-01 #8602 @elbeejay
@tennlee will you consider reviewing this submission titled: *Lenapy: Enhancing xarray for multidimensional geophysical data analysis for JOSS?* Given your JOSS publication on `scores`, and experience with `xarray`, I thought you might be able to provide a well-thought out and constructive review.
