### 2025-10-06 cd2789d
refact: improve climato.signal
