### 2025-10-12 #9189 @editorialbot
[REVIEW]: Lenapy: Enhancing xarray for multidimensional geophysical data analysis
**Submitting author:** @hulecom (<a href="http://orcid.org/0000-0002-7007-4748">Hugo Lecomte</a>)
**Repository:** https://github.com/CNES/lenapy
**Branch with paper.md** (empty if default branch): joss
**Version:** V1.0
