+++ paper.bib
@@ -97,7 +97,7 @@
   author = {Andreas Kvas},
   journal = {GitHub repository},
   publisher = {GitHub},
-  url = {https://https://github.com/akvas/grates},
+  url = {https://github.com/akvas/grates},
   year = {2024}
 }
 
