### 2025-10-30 51ed355
Remove setup.py and update pyproject.toml
