### 2025-10-30 7cf08ad
Change minimal python version according to the CI
