+++ pyproject.toml
@@ -11,7 +11,7 @@ authors = [
     { name = "Sebastien Fourest & EMC2 team", email = "sebastien.fourest@cnes.fr" }
 ]
 license = { text = "GPL-3.0" }
-requires-python = ">=3.7"
+requires-python = ">=3.11"
 dependencies = [
     "matplotlib>=3.6",
     "xesmf>=0.8.2",
