### 2025-10-30 f9b4c80
Change minimal python version according to the CI
