### 2025-10-30 995ae46
Update rename_data doc after removal of setup.py
