### 2025-10-31 c43fc3c
Update README.md with more information
