+++ lenapy/plots/plot_trend.py
@@ -1,6 +1,7 @@
 import numpy as np
-import pandas as od
+import pandas as pd
 import xarray as xr
+import matplotlib.pyplot as plt
 
 
 def plot_trend(data, sigma, out="OLS", freq="1M", save=None, **kwargs):
