### 2025-11-03 CNES/lenapy#10 @miniufo
Notebooks not reproducible
This is part of the review  for JOSS [#9189](https://github.com/openjournals/joss-reviews/issues/9189).

I've tried to create the required environment using `conda env create -f environment.yml` where `environment.yml` is in the `lenapy` folder, and run the notebooks to see if they are reproducible.  The first one `Climato.ipynb` is fine.  But from the second one, it reports an error:

I just go through the docs and find no information about how to add an (extra) engine **lenapyNetcdf** for this dataset.  Is this extra engine already in the source code or should I need to install it from somewhere else?
