### 2025-11-03 CNES/lenapy#11 @miniufo
add some notes on the docs
This is part of the review for JOSS [#9189](https://github.com/openjournals/joss-reviews/issues/9189).

The docs and tests are great.  There is only one thing that when I first went through the welcome page of the documentation, I saw the four modules:
* lenapy_ocean: encapsulation of the GSW library containing the main ocean thermodynamic calculation routines (TEOS10)
* lenapy_geo: current operations on gridded data (longitude/latitude/depth)
* lenapy_time: common operations on time series (filtering, climatology, regressions, etc.)
* lenapy_harmo: operations related to spherical harmonics decomposition and their projections on latitude/longitude grids

I didn't realize first that the four modules are renamed into `lnocean`, `lngeo`, `lntime`, and `lnharmo` as data accessors.  So it would be great to add this abbrev. module names on the docs, so that potential users may relate the names to the modules.

Also, the docs link can be put on the upper-right corner of the repo, with a sentence of description at:

where currently it is blank.
