### 2025-11-03 CNES/lenapy#12 @miniufo
add a state of the field in the paper
This is part of the review for JOSS [#9189](https://github.com/openjournals/joss-reviews/issues/9189)

There is an item on my checklist
- State of the field: Do the authors describe how this software compares to other commonly-used packages?

which is not included in the present JOSS manuscript.  I think some related contents are already included in the **Statement of need** section, as the authors mentioned several related packages like `pyshtools` and `grates`.  It would be much better to organize them into a single section as required, with some extra efforts to include:
- [xeofs](https://github.com/xarray-contrib/xeofs)
- [windspharm](https://github.com/ajdawson/windspharm)
- [xclim](https://github.com/Ouranosinc/xclim)

which are all based on `xarray`'s ecosystem and are related to the current submodules.  For potential users, it would be good to know the differences between them and lenapy, and therefore help them choose the right tools.
