### 2025-11-03 #9189 @miniufo
Sorry for a bit late.  Just go through the paper and the repo.  It is a nice package for potential users in our field.  I've opened several issues so that the authors can address them accordingly.  Two things related to my checklist are:
- I find two contributors but three authors on the JOSS paper.  Just make sure this is OK @elbeejay ;
- The paper lacks a section on "State of the field", where more related packages should be mentioned and compared in this section (see [issue #12](https://github.com/CNES/lenapy/issues/12)).

You can address my issues through the opened ones (not here).  After all of the issues are closed, we may then return to the checklist to see if everything is OK.
