### 2025-11-07 9dd912a
Add CITATION.cff
