### 2025-11-10 CNES/lenapy#10 @hulecom
For tutorial_nharmo.ipynb, there is indeed a problem from my side

I added the 'Apply simple spatial filtering' part without considering that it will modify the rest of the notebook
So I corrected it by just creating a new dataset **ds_filtered** instead of replacing ds

This correct the whole thing
