+++ paper.md
@@ -31,7 +31,7 @@ bibliography: paper.bib
 # Summary
 
 `Lenapy` is a Python library designed to facilitate the processing and analysis of geophysical and climate datasets, such as those used in oceanography, geodesy, and Earth observation in general. 
-Built on top of `xarray` and fully compatible with `dask`, it enables scalable workflows on multidimensional datasets using community standards for data formats (NetCDF) and metadata conventions (CF).
+Built on top of `xarray` and fully compatible with `Dask`, it enables scalable workflows on multidimensional datasets using community standards for data formats (NetCDF) and metadata conventions (CF).
 
 `Lenapy` provides high-level accessors that extend `xarray.Dataset` and `xarray.DataArray` objects, allowing application of specialized methods.
 
@@ -48,31 +48,33 @@ Analyzing these data requires specialized operations such as harmonic analysis, 
 For the oceanography part, the accurate computation of seawater properties is essential in physical oceanography.
 The Thermodynamic Equation of Seawater (TEOS-10) framework provides consistent definitions and algorithms for quantities such as potential temperature, conservative temperature, and density [@Mcdougall_2011].
 
-Existing libraries address specific aspects of these requirements:
-
-- `pyshtools` [@Wieczorek_2018] is a comprehensive package for spherical harmonic transforms and spectral analysis, particularly in geophysics. Yet, it operates on standalone arrays and does not natively support xarray, limiting its compatibility with NetCDF-based workflows.
-- `grates` [@Kvas] provides object-oriented tools for spherical harmonics, but similarly only operates on standalone arrays.
-- `gsw-xarray` [@Caneill_2024] provides implementations of the TEOS-10, facilitating oceanographic computations. It offers a wrapper around `GSW-Python` for xarray objects, but without Dask support.
-
-While `gsw-xarray` is more complete than our GSW wrapper in `Lenapy`, our library propose complementary geodetic tools for spatial and spherical harmonics operations.
-
-To our knowledge, no other existing Python library offers a coherent suite of oceanographic and geophysical operations (detailed below) within a unified, xarray-native framework supporting both scalability (via Dask [@Dask_2016]) and labeled, multidimensional arrays.
-Moreover, critical geospatial utilities (such as surface-aware averaging, weighted statistics, spherical distance computation, and climatology fitting) remain fragmented across ecosystems or require custom implementations.
-
-`Lenapy` addresses this gap by providing a modular Python package built on xarray and Dask, exposing accessors (.lngeo, .lnharmo, .lnocean, .lntime) for direct and simple application of domain-specific methods to xarray.Dataset or xarray.DataArray objects. 
-For example, users can compute area-weighted means via `ds.lngeo.mean()` or extract the global ocean heat content using `ds.lnocean.gohc()`.
-
 `Lenapy` is designed for Earth scientists, oceanographers, climate researchers, and geodesists who routinely manipulate global or regional gridded datasets and require specific processing workflows.
 `Lenapy` aims to maintain full compatibility with the PyData ecosystem.
 
 Furthermore, `Lenapy` offers a unified approach for calculating Global Mean Sea Level by integrating both steric and manometric components, as well as relative sea-level changes [@Gregory_2019].
 `Lenapy` facilitates this decomposition with a unique Python library that compute these components directly from xarray-based datasets, enabling researchers to analyze sea-level changes comprehensively within a consistent framework.
 
+# State of the field
+Existing Python libraries address specific needs in geophysical and climate data analysis, but often in a fragmented or domain-specific manner.
+`pyshtools` [@Wieczorek_2018] and `grates` [@Kvas] offer tools for gravity spherical harmonic analysis.
[…]
