### 2025-11-11 7f8c3bf
Update State of the field part according to comments
