+++ paper.bib
@@ -77,28 +77,6 @@
   doi = {10.1029/2018GC007529},
 }
 
-@article{Rieger_2024,
-author = {Rieger, Niclas and Levang, Samuel J.},
-doi = {10.21105/joss.06060},
-journal = {Journal of Open Source Software},
-number = {93},
-pages = {6060},
-title = {{xeofs: Comprehensive EOF analysis in Python with xarray}},
-volume = {9},
-year = {2024}
-}
-
-@article{Bourgault_2023,
-author = {Bourgault, Pascal and Huard, David and Smith, Trevor James and Logan, Travis and Aoun, Abel and Lavoie, Juliette and Dupuis, Éric and Rondeau-Genesse, Gabriel and Alegre, Raquel and Barnes, Clair and Laperrière, Alexis Beaupré and Biner, Sébastien and Caron, David and Ehbrecht, Carsten and Fyke, Jeremy and Keel, Tom and Labonté, Marie-Pier and Lierhammer, Ludwig and Low, Jwen-Fai and Quinn, Jamie and Roy, Philippe and Squire, Dougie and Stephens, Ag and Tanguy, Maliko and Whelan, Christopher},
-doi = {10.21105/joss.05415},
-journal = {Journal of Open Source Software},
-number = {85},
-pages = {5415},
-title = {{xclim: xarray-based climate data analytics}},
-volume = {8},
-year = {2023}
-}
-
 @article{Dawson_2016,
   title = {Windspharm: A High-Level Library for Global Wind Field Computations Using Spherical Harmonics},
   author = {Dawson, Andrew},
