### 2025-11-13 e15625c
sort imports in plot_trend.py
