### 2025-11-26 #9189 @hulecom
On my side, the re-generated version of the draft seems OK.
