### 2025-11-29 #9189 @elbeejay
Yes we shall move on to the final stages of the JOSS process. Many thanks to our volunteer peer-reviewers, @ThomasMGeo and @miniufo, for their time, effort, and responsiveness. JOSS would not be possible without your efforts.

@hulecom I will be generating a post-review checklist where we will finalize some of the metadata associated with your paper. You will not be able to modify the checklist directly, so please simply post here when you have completed the various tasks.
