### 2025-12-02 6c303d9
New version for JOSS publication
