+++ lenapy/__init__.py
@@ -1,10 +1,18 @@
+from importlib.metadata import version as _version
 import warnings
 
 from lenapy import constants, lenapy_geo, lenapy_harmo, lenapy_time
 
 try:
+    __version__ = _version("lenapy")
+except Exception:
+    # Local copy or not installed with setuptools.
+    # Disable minimum version checks on downstream libraries.
+    __version__ = "9999"
+
+try:
     from lenapy import lenapy_ocean
-except:
+except Exception:
     warnings.warn("To use lenapy_ocean, please install gsw : pip install gsw>=3.6.16")
 import cf_xarray as cfxr
 
