+++ pyproject.toml
@@ -4,11 +4,11 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "lenapy"
-version = "1.0"
-description = "First official release of lenapy on github"
+version = "1.1"
+description = "Geophysical analysis in Python with xarray"
 readme = "README.md"
 authors = [
-    { name = "Sebastien Fourest & EMC2 team", email = "sebastien.fourest@cnes.fr" }
+    { name = "Sebastien Fourest & Hugo Lecomte & EMC2 team", email = "sebastien.fourest@cnes.fr" }
 ]
 license = { text = "GPL-3.0" }
 requires-python = ">=3.11"
