### 2025-12-02 6137ed5
Modif init
