### 2025-12-02 4cbff9d
New affiliation (post-doc change) for Hugo Lecomte
