### 2025-12-03 02fa419
Citation for grates
