+++ paper.bib
@@ -103,13 +103,12 @@
   url = {https://dask.org}
 }
 
-@misc{Kvas,
-  title = {grates},
+@software{Kvas,
+  title = {akvas/grates: v0.1.0},
   author = {Andreas Kvas},
-  journal = {GitHub repository},
-  publisher = {GitHub},
-  url = {https://github.com/akvas/grates},
-  year = {2024}
+  year = {2024},
+  doi = {10.5281/zenodo.11672421},
+  howpublished = {Zenodo}
 }
 
 @software{Caneill_2024,
