### 2025-12-03 #9189 @hulecom
Hello @elbeejay

I have modified my affiliation to include my new post-doc affiliation since August (given the time I spent for recent reviews correction and finalizing this publication)

We did a new release of the code with an up to date version including correction from reviews `V1.1`

The Zenodo archive is [https://doi.org/10.5281/zenodo.17801405](https://doi.org/10.5281/zenodo.17801405)

For the skip DOIs, I found the zenodo archive for akvas/grates code (updated on the joss branch of the repo). The rest of skip DOIs are PhD manuscript, book or manual stored without DOI.

Hope I didn't miss anything
