### 2025-12-11 eff767a
Normalize titles in bibliography entries
