+++ paper.bib
@@ -7,7 +7,7 @@
 }
 
 @phdthesis{Dieng_2017,
-  title = {Variations Actuelles Du Niveau de La Mer},
+  title = {Variations Actuelles Du Niveau de la Mer},
   author = {Dieng, Habib Boubacar},
   year = {2017},
   address = {LEGOS},
@@ -15,7 +15,7 @@
 }
 
 @phdthesis{Meyssignac_2012,
-  title = {La Variabilit{\'e} R{\'e}gionale Du Niveau de La Mer},
+  title = {La Variabilit{\'e} R{\'e}gionale Du Niveau de la Mer},
   author = {Meyssignac, Benoit},
   year = {2012},
   address = {LEGOS},
@@ -23,7 +23,7 @@
 }
 
 @article{Bouih_2025,
-  title = {Regional Sea Level Budget over 2004-2022},
+  title = {Regional {Sea} {Level} {Budget} over 2004-2022},
   author = {Bouih, Marie and Barnoud, Anne and Yang, Chunxue and Storto, Andrea and Blazquez, Alejandro and Llovel, William and Fraudeau, Robin and Cazenave, Anny},
   year = {2025},
   journal = {EGUsphere},
@@ -33,7 +33,7 @@
 }
 
 @article{Blazquez_2018,
-  title = {Exploring the Uncertainty in {{GRACE}} Estimates of the Mass Redistributions at the {{Earth}} Surface: Implications for the Global Water and Sea Level Budgets},
+  title = {Exploring the {Uncertainty} in {{GRACE}} {Estimates} of the {Mass} {Redistributions} at the {{Earth}} {Surface}: {Implications} for the {Global} {Water} and {Sea} {Level} {Budgets}},
   author = {Blazquez, A and Meyssignac, B and {Lemoine Jean-Michel}, J.M. and Berthier, E and Ribes, A and Cazenave, A},
   year = {2018},
   journal = {Geophysical Journal International},
@@ -58,7 +58,7 @@
 
 @book{Mcdougall_2011,
   author = {McDougall, T. J. and Barker, P. M.},
-  title = {Getting Started with TEOS-10 and the Gibbs Seawater (GSW) Oceanographic Toolbox},
+  title = {{Getting} {Started} with {TEOS-10} and the {Gibbs} {Seawater} ({GSW}) {Oceanographic} {Toolbox}},
   year = {2011},
   pages = {28},
   publisher = {SCOR/IAPSO WG127},
@@ -78,7 +78,7 @@
 }
 
 @article{Dawson_2016,
-  title = {Windspharm: A High-Level Library for Global Wind Field Computations Using Spherical Harmonics},
+  title = {Windspharm: {A} {High}-{Level} {Library} {for} {Global} {Wind} {Field} {Computations} {Using} {Spherical} {Harmonics}},
   author = {Dawson, Andrew},
   year = {2016},
   journal = {Journal of Open Research Software},
@@ -97,8 +97,8 @@
 }
 
 @manual{Dask_2016,
-  title = {Dask: Library for dynamic task scheduling},
-  author = {Dask Development Team},
+  title = {{Dask}: {Library} for dynamic task scheduling},
+  author = {{Dask} {Development} {Team}},
   year = {2016},
   url = {https://dask.org}
 }
@@ -112,7 +112,7 @@
 }
 
 @software{Caneill_2024,
-  title = {Gsw-Xarray},
+  title = {gsw-xarray},
   author = {Caneill, Romain and Barna, Andrew},
   year = {2024},
   doi = {10.5281/zenodo.11382921},
