### 2025-12-11 2dbcac8
Fix title encoding in Blazquez 2020 thesis entry
