+++ paper.bib
@@ -1,5 +1,5 @@
 @phdthesis{Blazquez_2020,
-  title = {Caract{\'e}risation Par Satellite Des {\'E}changes d'eau Entre l'oc{\'e}an et Les Continents Aux {\'E}chelles Interannuelles {\`a} D{\'e}cennales},
+  title = {Caract{\'e}risation Par Satellite Des échanges d'eau Entre l'oc{\'e}an et Les Continents Aux échelles Interannuelles {\`a} D{\'e}cennales},
   author = {Blazquez, Alejandro},
   year = {2020},
   address = {LEGOS},
