### 2025-12-11 6b3d637
Fix formatting of software titles in paper.bib
