+++ paper.bib
@@ -104,7 +104,7 @@
 }
 
 @software{Kvas,
-  title = {akvas/grates: v0.1.0},
+  title = {{akvas/grates}: v0.1.0},
   author = {Andreas Kvas},
   year = {2024},
   doi = {10.5281/zenodo.11672421},
@@ -112,7 +112,7 @@
 }
 
 @software{Caneill_2024,
-  title = {gsw-xarray},
+  title = {{gsw}-xarray},
   author = {Caneill, Romain and Barna, Andrew},
   year = {2024},
   doi = {10.5281/zenodo.11382921},
