### 2025-12-15 #9189 @hulecom
Thank you @kthyng and thank you to everyone @elbeejay, @ThomasMGeo, @miniufo for your comments, suggestions and the overall help on the article. This was my first publication with JOSS and the whole process was really interesting !
