### 2026-01-26 befcf21
Debug gravi reader for static file EIGEN-6C4
