### 2026-01-26 e2feb94
Add dask gestion to grid_to_sh and sh_to_grid
