+++ lenapy/utils/harmo.py
@@ -238,7 +238,7 @@ def _init_degrees(
     mmax: int = None,
     used_l: np.ndarray = None,
     used_m: np.ndarray = None,
-) -> tuple[np.ndarray, np.ndarray, int, int, int, int]:
+) -> tuple[xr.Dataset, np.ndarray, np.ndarray, int, int, int, int]:
     """
     Initialize spherical harmonic degrees and orders to be used.
 
@@ -257,6 +257,8 @@ def _init_degrees(
 
     Returns
     -------
+    sub_data : xr.Dataset
+        Reduced dataset with selected degrees and orders.
     used_l: np.ndarray
         Degrees to use.
     used_m: np.ndarray
@@ -276,7 +278,8 @@ def _init_degrees(
     mmin = int(data.m.min()) if mmin is None else mmin
     used_l = np.arange(lmin, lmax + 1) if used_l is None else used_l
     used_m = np.arange(mmin, mmax + 1) if used_m is None else used_m
-    return used_l, used_m, lmin, lmax, mmin, mmax
+    sub_data = data.sel(l=used_l, m=used_m)
+    return sub_data, used_l, used_m, lmin, lmax, mmin, mmax
 
 
 def sh_to_grid(
@@ -304,6 +307,8 @@ def sh_to_grid(
     include_elastic: bool = True,
     plm: xr.DataArray = None,
     normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
+    use_dask: bool = False,
+    chunks_plm: dict | None = None,
     **kwargs,
 ) -> xr.DataArray:
     """
@@ -374,6 +379,11 @@ def sh_to_grid(
         Either '4pi', 'ortho', or 'schmidt' for 4pi normalized, orthonormalized, or Schmidt semi-normalized SH
         functions, respectively. Default is '4pi'.
 
+    use_dask : bool, optional
+        If True, use dask to chunk plm for memory optimization. Default is False.
+    chunks_plm : dict, optional
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 1}.
+
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
         for the unit conversion. These parameters include (see :func:`l_factor_conv` documentation for more details) :
@@ -386,7 +396,7 @@ def sh_to_grid(
     """
     # add mask in output variable
 
-    used_l, used_m, lmin, lmax, mmin, mmax = _init_degrees(
+    sub_data, used_l, used_m, lmin, lmax, mmin, mmax = _init_degrees(
         data, lmin, lmax, mmin, mmax, used_l, used_m
     )
     used_l, use_czero_coef, force_mass_conservation = _handle_mass_conservation(
@@ -415,22 +425,22 @@ def sh_to_grid(
             plm = compute_plm(
                 lmax,
                 np.cos(geocentric_colat),
+                latitude=latitude,
                 mmax=mmax,
                 normalization=normalization_plm,
+                use_dask=use_dask,
+                chunks=chunks_plm,
             )
         else:
             plm = compute_plm(
-                lmax, sin_latitude, mmax=mmax, normalization=normalization_plm
+                lmax,
+                sin_latitude,
+                latitude=latitude,
+                mmax=mmax,
+                normalization=normalization_plm,
+                use_dask=use_dask,
+                chunks=chunks_plm,
             )
-        plm = xr.DataArray(
-            plm,
-            dims=["l", "m", "latitude"],
-            coords={
-                "l": np.arange(lmax + 1),
-                "m": np.arange(mmax + 1),
-                "latitude": latitude,
[…]
