### 2026-01-26 3ce1495
Debug gravi readers
