### 2026-01-27 e35de09
Memory improvements
