### 2026-02-04 0343cfd
Change the lmax warning value due to futher test
