### 2026-02-04 e607461
Change apply parameter behavior because of the copy(deep=True) not acting as it should
