### 2026-02-05 de0e044
Debug gravi reader for ESM ESA files
