### 2026-02-09 2d3b18c
try to minimize numerical error for normal zonal correction
