### 2026-02-09 d944c12
Test plm for lmax=2190
