### 2026-02-10 59405db
Update plm function to return first derivative with associated argument. Reorganize harmo.py plm functions
