+++ lenapy/utils/harmo.py
@@ -17,7 +17,7 @@ import datetime
 import inspect
 import pathlib
 import warnings
-from typing import Literal
+from typing import Literal, Optional
 
 import cf_xarray as cfxr
 import numpy as np
@@ -27,84 +27,6 @@ import xarray as xr
 
 from lenapy.constants import *
 from lenapy.utils.gravity import estimate_normal_gravity
-
-
-def _compute_factors(
-    lmax: int, normalization: Literal["4pi", "ortho", "schmidt"]
-) -> tuple[np.ndarray, np.ndarray, float, float]:
-    """
-    Compute recurrence coefficients f1 and f2 for the Legendre recursion.
-
-    Parameters
-    ----------
-    lmax : int
-        Maximum degree.
-    normalization : {'4pi', 'ortho', 'schmidt'}
-        Normalization scheme.
-
-    Returns
-    -------
-    f1 : np.ndarray
-        Recurrence factor f1.
-    f2 : np.ndarray
-        Recurrence factor f2.
-    norm_p10 : float
-        Normalization for P(1,0).
-    norm_4pi : float
-        Overall normalization factor.
-    """
-    size = (lmax + 1) * (lmax + 2) // 2
-    f1 = np.zeros(size)
-    f2 = np.zeros(size)
-
-    k = 2
-    if normalization in ("4pi", "ortho"):
-        norm_p10 = np.sqrt(3)
-        norm_4pi = 1 if normalization == "4pi" else 4 * np.pi
-        for l in range(2, lmax + 1):
-            k += 1
-            f1[k] = np.sqrt(2 * l - 1) * np.sqrt(2 * l + 1) / l
-            f2[k] = (l - 1) * np.sqrt(2 * l + 1) / (np.sqrt(2 * l - 3) * l)
-            for m in range(1, l - 1):
-                k += 1
-                f1[k] = (
-                    np.sqrt(2 * l + 1)
-                    * np.sqrt(2 * l - 1)
-                    / (np.sqrt(l + m) * np.sqrt(l - m))
-                )
-                f2[k] = (
-                    np.sqrt(2 * l + 1)
-                    * np.sqrt(l - m - 1)
-                    * np.sqrt(l + m - 1)
-                    / (np.sqrt(2 * l - 3) * np.sqrt(l + m) * np.sqrt(l - m))
-                )
-            k += 2
-    elif normalization == "schmidt":
-        norm_p10 = 1
-        norm_4pi = 1
-        for l in range(2, lmax + 1):
-            k += 1
-            f1[k] = (2 * l - 1) / l
-            f2[k] = (l - 1) / l
-            for m in range(1, l - 1):
-                k += 1
-                f1[k] = (2 * l - 1) / (np.sqrt(l + m) * np.sqrt(l - m))
-                f2[k] = (
-                    np.sqrt(l - m - 1)
-                    * np.sqrt(l + m - 1)
-                    / (np.sqrt(l + m) * np.sqrt(l - m))
-                )
-            k += 2
-    else:
-        raise ValueError(
-            (
-                f"Unknown normalization '{normalization}'. "
-                "It should be either "
-                "'4pi', 'ortho' or 'schmidt'"
-            )
-        )
-
-    return f1, f2, norm_p10, norm_4pi
 
 
 def _generate_grid(
@@ -443,28 +365,7 @@ def sh_to_grid(
             )
 
     else:
-        # Verify plm integrity
-        if (
-            not isinstance(plm, xr.DataArray)
-            or "l" not in plm.coords
-            or "m" not in plm.coords
-            or "latitude" not in plm.coords
-        ):
-            raise TypeError(
[…]
