### 2026-02-10 d43c46b
typos of harmo.py
