### 2026-02-11 ce87cc2
refactor apply_normal_zonal_correction
