### 2026-02-11 117c438
Gravity distrubance estimation and function refactor. To be tested
