+++ lenapy/utils/harmo.py
@@ -26,6 +26,7 @@ import scipy as sc
 import xarray as xr
 
 from lenapy.constants import *
+from lenapy.utils.geo import latitude_to_geocentric_colatitude
 
 
 def _generate_grid(
@@ -205,7 +206,9 @@ def _init_degrees(
 
 def sh_to_grid(
     data: xr.Dataset,
-    unit: Literal["mewh", "mmgeoid", "microGal", "bar", "mvcu", "norm"] = "mewh",
+    unit: Literal[
+        "mewh", "mmgeoid", "microGal", "potential", "pascal", "mvcu", "norm"
+    ] = "mewh",
     errors=False,
     lmax: int | None = None,
     mmax: int | None = None,
@@ -244,7 +247,7 @@ def sh_to_grid(
     data : xr.Dataset
         xr.Dataset that corresponds to SH data to convert into spatial representation.
     unit : str, optional
-        'mewh', 'mmgeoid', 'microGal', 'bar', 'mvcu', or 'norm'
+        'mewh', 'mmgeoid', 'microGal', 'potential', 'pascal', 'mvcu', or 'norm'
         Unit of the spatial data used in the transformation. Default is 'mewh' for meters of Equivalent Water Height.
         See utils.harmo.l_factor_conv() doc for details on the units.
     errors : bool, optional
@@ -285,11 +288,11 @@ def sh_to_grid(
         List of latitude to use for the grid computation (if given, others latitude information are not considered).
 
     force_mass_conservation : bool, optional
-        If True, force that the grid resulting from all coefficients except C0 has a null global mass. Default is False
+        If True, force that the grid resulting from all coefficients except C0 has a null global mass. Default is False.
     ellipsoidal_earth : bool, optional
         If True, consider the Earth as an ellipsoid following [Ditmar2018]. Default is False for a spherical Earth.
     include_elastic : bool, optional
-        If True, the Earth behavior is elastic. Default is True
+        If True, the Earth behavior is elastic. Default is True.
 
     plm : xr.DataArray, optional
         Precomputed plm values as a xr.DataArray variable. For example with the code :
@@ -315,8 +318,7 @@ def sh_to_grid(
     xgrid : xr.DataArray
         Spatial representation of the SH Dataset in the chosen unit.
     """
-    # add mask in output variable
-
+    # Get parameters for computation of the grid and for the unit conversion
     sub_data, used_l, used_m, lmin, lmax, mmin, mmax = _init_degrees(
         data, lmin, lmax, mmin, mmax, used_l, used_m
     )
@@ -330,38 +332,22 @@ def sh_to_grid(
         bounds, dlon, dlat, longitude, latitude, radians_in
     )
 
-    cos_latitude = np.cos(np.deg2rad(latitude))
-    sin_latitude = np.sin(np.deg2rad(latitude))
-    f_earth = kwargs["f_earth"] if "f_earth" in kwargs else LNPY_F_EARTH_GRS80
-    geocentric_colat = xr.DataArray(
-        np.arctan2(cos_latitude, (1 - f_earth) ** 2 * sin_latitude),
-        dims=["latitude"],
-        coords={"latitude": latitude},
+    geocentric_colat = latitude_to_geocentric_colatitude(
+        latitude, ellipsoidal_earth=ellipsoidal_earth, **kwargs
     )
 
     # -- beginning of computation
     # Computing plm for converting to spatial domain
     if plm is None:
-        if ellipsoidal_earth:
-            plm = compute_plm(
-                lmax,
-                np.cos(geocentric_colat),
-                latitude=latitude,
[…]
