+++ lenapy/utils/gravity.py
@@ -34,6 +34,15 @@ import numpy as np
 import xarray as xr
 
 from lenapy.constants import *
+from lenapy.utils.geo import latitude_to_geocentric_colatitude
+from lenapy.utils.harmo import (
+    _assert_plm,
+    _generate_grid,
+    _get_earth_parameters,
+    _init_bounds_and_grid,
+    compute_plm,
+    l_factor_conv,
+)
 
 
 def change_reference(
@@ -374,9 +383,9 @@ def estimate_normal_gravity(
     )
 
 
-def apply_normal_zonal_correction(
+def normal_zonal_correction(
     ds: xr.Dataset,
-    radius: float | None = None,
+    a_earth: float | None = None,
     earth_gravity_constant: float | None = None,
     f_earth: float = LNPY_F_EARTH_GRS80,
     omega_earth: float = LNPY_OMEGA_EARTH_GRS80,
@@ -384,13 +393,13 @@ def apply_normal_zonal_correction(
     apply: bool = True,
 ) -> xr.Dataset:
     """
-    Apply a correction of the normal gravity field on zonal coefficients on a SH dataset for a specified ellipsoid.
+    Normal gravity field correction on zonal coefficients on a SH dataset for a specified ellipsoid.
 
     Parameters
     ----------
     ds : xr.Dataset
         xr.Dataset that corresponds to SH data to be correct for the normal gravity field.
-    radius : float | None, optional
+    a_earth : float | None, optional
         Earth radius constant of the dataset ds in meters. If not provided, uses `ds.attrs['radius']`.
     earth_gravity_constant : float | None, optional
         Current gravitational constant of the Earth of the dataset ds in m³/s².
@@ -416,7 +425,7 @@ def apply_normal_zonal_correction(
 
     """
     try:
-        radius = ds.attrs["radius"] if radius is None else radius
+        a_earth = ds.attrs["radius"] if a_earth is None else a_earth
         earth_gravity_constant = (
             ds.attrs["earth_gravity_constant"]
             if earth_gravity_constant is None
@@ -446,7 +455,7 @@ def apply_normal_zonal_correction(
         q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
 
         k = 1 / 3 - (
-            2 * omega_earth**2 * radius**3 * np.sqrt(2 * f_earth - f_earth**2)
+            2 * omega_earth**2 * a_earth**3 * np.sqrt(2 * f_earth - f_earth**2)
         ) / (45 * earth_gravity_constant * q0)
 
         correction = (
@@ -464,6 +473,480 @@ def apply_normal_zonal_correction(
     )
 
     return ds_out
+
+
+def sh_to_gravity_disturbance(
+    data: xr.Dataset,
+    lonmin: float = -180,
+    lonmax: float = 180,
+    latmin: float = -90,
+    latmax: float = 90,
+    bounds: list = None,
+    dlon: float = 1,
+    dlat: float = 1,
+    longitude: np.ndarray | None = None,
+    latitude: np.ndarray | None = None,
+    radians_in: bool = False,
+    ellipsoidal_earth: bool = False,
+    normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
+    use_dask: bool = False,
+    chunks_plm: dict | None = None,
+    **kwargs,
+) -> xr.DataArray:
+    """
+    Transform Spherical Harmonics (SH) dataset into a grid corresponding to the gravity disturbance.
+    With choice for constants, unit, love numbers, degree/order, spatial grid latitude and longitude, Earth hypothesis.
+
+    Parameters
+    ----------
+    data : xr.Dataset
[…]
