### 2026-02-23 da90d1b
Advancement on gravity disturbance and SH logics
