+++ lenapy/utils/gravity.py
@@ -34,12 +34,11 @@ import numpy as np
 import xarray as xr
 
 from lenapy.constants import *
-from lenapy.utils.geo import latitude_to_geocentric_colatitude
+from lenapy.utils.geo import assert_latitude, latitude_to_geocentric_colatitude
 from lenapy.utils.harmo import (
     _assert_plm,
     _generate_grid,
     _get_earth_parameters,
-    _init_bounds_and_grid,
     compute_plm,
     l_factor_conv,
 )
@@ -477,6 +476,7 @@ def normal_zonal_correction(
 
 def sh_to_gravity_disturbance(
     data: xr.Dataset,
+    surface: Literal[None, "reference", "geoid"] | float | xr.DataArray = None,
     lonmin: float = -180,
     lonmax: float = 180,
     latmin: float = -90,
@@ -490,17 +490,24 @@ def sh_to_gravity_disturbance(
     ellipsoidal_earth: bool = False,
     normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
     use_dask: bool = False,
+    chunks_lfactor: dict | None = None,
     chunks_plm: dict | None = None,
+    test=True,
     **kwargs,
 ) -> xr.DataArray:
     """
     Transform Spherical Harmonics (SH) dataset into a grid corresponding to the gravity disturbance.
+    This grid can be the reference surface (sphere or ellipsoid) or another surface.
     With choice for constants, unit, love numbers, degree/order, spatial grid latitude and longitude, Earth hypothesis.
 
     Parameters
     ----------
     data : xr.Dataset
         xr.Dataset that corresponds to SH data to convert into spatial representation.
+    surface : float | xr.DataArray | Literal[None, "reference", "geoid"], optional
+        Surface to consider for the computation of the grid. If None or "reference", consider the reference sphere or
+        ellipsoid. If "geoid", consider the geoid as surface computed from the given field.
+        If a float or xr.DataArray is given, consider it as the surface height in meter to add to the reference surface.
 
     lonmin : float, optional
         Minimal longitude of the future grid.
@@ -538,6 +545,8 @@ def sh_to_gravity_disturbance(
 
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
+    chunks_lfactor : dict, optional
+        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 20}.
     chunks_plm : dict, optional
         Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 1}.
 
@@ -552,12 +561,17 @@ def sh_to_gravity_disturbance(
         Spatial representation of the SH Dataset in gravity disturbance.
     """
     # Get parameters for computation of the grid and for the unit conversion
-    bounds, dlon, dlat = _init_bounds_and_grid(
-        bounds, lonmin, lonmax, latmin, latmax, dlon, dlat, radians_in
-    )
-
     longitude, latitude = _generate_grid(
-        bounds, dlon, dlat, longitude, latitude, radians_in
+        bounds,
+        lonmin,
+        lonmax,
+        latmin,
+        latmax,
+        dlon,
+        dlat,
+        radians_in,
+        longitude,
+        latitude,
     )
 
     data_null = xr.zeros_like(data)
@@ -589,89 +603,11 @@ def sh_to_gravity_disturbance(
     )
 
     # -- beginning of computation
-    ds_centrifugal = centrifugal_potential_partial_derivative(
-        latitude, ellipsoidal_earth, **kwargs
-    )
-
[…]
