+++ lenapy/utils/harmo.py
@@ -26,55 +26,14 @@ import scipy as sc
 import xarray as xr
 
 from lenapy.constants import *
-from lenapy.utils.geo import latitude_to_geocentric_colatitude
+from lenapy.utils.geo import (
+    assert_grid,
+    assert_latitude,
+    latitude_to_geocentric_colatitude,
+)
 
 
 def _generate_grid(
-    bounds: list[float],
-    dlon: float,
-    dlat: float,
-    longitude: np.ndarray | None,
-    latitude: np.ndarray | None,
-    radians_in: bool,
-) -> tuple[np.ndarray, np.ndarray]:
-    """
-    Generate longitude and latitude arrays for the grid.
-
-    load longitude and latitude if given
-    if not : compute longitude and latitude in degrees between given or defaults bounds
-
-    Parameters
-    ----------
-    bounds : list of float
-        [lonmin, lonmax, latmin, latmax]
-    dlon, dlat : float
-        Grid spacing in degrees.
-    longitude, latitude : np.ndarray or None
-        Optionally provided coordinate arrays.
-    radians_in : bool
-        Whether the input arrays are in radians.
-
-    Returns
-    -------
-    longitude : np.ndarray
-        Array of longitudes in degrees.
-    latitude : np.ndarray
-        Array of latitudes in degrees.
-    """
-    if longitude is None:
-        longitude = np.arange(bounds[0] + dlon / 2.0, bounds[1] + dlon / 2.0, dlon)
-    elif radians_in:
-        longitude = np.rad2deg(longitude)
-
-    if latitude is None:
-        latitude = np.arange(bounds[2] + dlat / 2.0, bounds[3] + dlat / 2.0, dlat)
-    elif radians_in:
-        latitude = np.rad2deg(latitude)
-
-    return longitude, latitude
-
-
-def _init_bounds_and_grid(
     bounds: list[float] | None,
     lonmin: float,
     lonmax: float,
@@ -83,9 +42,16 @@ def _init_bounds_and_grid(
     dlon: float,
     dlat: float,
     radians_in: bool,
-) -> tuple[list[float], float, float]:
+    longitude: np.ndarray | None = None,
+    latitude: np.ndarray | None = None,
+    grid: xr.DataArray | xr.Dataset | None = None,
+) -> tuple[np.ndarray, np.ndarray]:
     """
     Initialize bounds and grid resolution.
+    Generate longitude and latitude arrays for the grid.
+
+    Load longitude and latitude from the given grid, if not from the given latitude, longitude and
+    if not : compute longitude and latitude in degrees between given or defaults bounds
 
     Parameters
     ----------
@@ -97,15 +63,19 @@ def _init_bounds_and_grid(
         Grid resolution in degrees or radians.
     radians_in: bool
         Whether the inputs are in radians.
+    longitude, latitude: np.ndarray or None
+        Optionally provided coordinate arrays.
+    radians_in: bool
+        Whether the input arrays are in radians.
+    grid: xr.DataArray | xr.Dataset | float
+        Optionally provided grid from which to extract longitude and latitude coordinates.
 
     Returns
     -------
-    bounds: list of float
-        Bounds in degrees.
-    dlon: float
-        Longitude resolution in degrees.
-    dlat: float
-        Latitude resolution in degrees.
+    longitude : np.ndarray
+        Array of longitudes in degrees.
+    latitude : np.ndarray
+        Array of latitudes in degrees.
     """
     if bounds is None:
         bounds = [lonmin, lonmax, latmin, latmax]
[…]
