### 2026-02-23 e2b5ea3
compute_plm can now use dask in the estimation to reduce maximal memory usage
