+++ lenapy/utils/harmo.py
@@ -308,7 +308,7 @@ def sh_to_grid(
     chunks_lfactor : dict, optional
         Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 20}.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 1}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
@@ -509,7 +509,7 @@ def grid_to_sh(
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 1}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
@@ -714,23 +714,22 @@ def _scale_plm_factors(
 
 
 def _compute_plm_vector(
-    p: np.ndarray,
-    dp: np.ndarray,
     z: np.ndarray,
     lmax: int,
     normalization: Literal["4pi", "ortho", "schmidt"],
     derivative: bool,
     dtype: complex | float | type[complex] | type[float],
+    f1,
+    f2,
+    norm_p10,
+    norm_4pi,
+    df,
 ) -> tuple[np.ndarray, np.ndarray]:
     """
     Compute associated Legendre functions P(l,m) and optionally their derivatives as a array.
 
     Parameters
     ----------
-    p : np.ndarray
-        Preallocated array for P(l,m) values.
-    dp : np.ndarray
-        Preallocated array for derivatives of P(l,m) values.
     lmax : int
         Maximum degree of legrendre functions.
     z : np.ndarray
@@ -742,8 +741,23 @@ def _compute_plm_vector(
         If True, compute the first derivative of the associated Legendre functions.
     dtype : dtype
         Data type of the output array.
+    f1 : np.ndarray
+        Recurrence factor f1.
+    f2 : np.ndarray
+        Recurrence factor f2.
+    norm_p10 : float
+        Normalization for P(1,0).
+    norm_4pi : float
+        Overall normalization factor.
+    df : np.ndarray | None
+        Recurrence factor for the derivative of plm, only returned if derivative is True.
     """
-    f1, f2, norm_p10, norm_4pi, df = _scale_plm_factors(lmax, normalization, derivative)
+    p = np.zeros(((lmax + 1) * (lmax + 2) // 2, len(z)), dtype=dtype)
+    dp = (
+        np.zeros(((lmax + 1) * (lmax + 2) // 2, len(z)), dtype=dtype)
+        if derivative
+        else None
+    )
 
     # scale factor based on Holmes2002
     scalef = 1e-280
@@ -827,6 +841,34 @@ def _compute_plm_vector(
     return p, dp
 
 
+def _compute_plm_ufunc(
+    z, lmax, mmax, normalization, derivative, dtype, f1, f2, norm_p10, norm_4pi, df
+):
+    """
+    Wrapper to use xr.apply_ufunc() to estimate plm with chunked latitudes.
+    Compute associated Legendre functions P(l,m) and optionally their derivatives as a array.
+
+    Parameters
+    ----------
+    see _compute_plm_vector Parameters list.
+    """
+    z = np.atleast_1d(np.asarray(z, dtype=dtype))
+
[…]
