### 2026-02-23 b843314
Change the unit gestion in l_factor_conv function with a new parser and a more flexible approach
