+++ lenapy/utils/harmo.py
@@ -201,9 +201,7 @@ def _init_degrees(
 
 def sh_to_grid(
     data: xr.Dataset,
-    unit: Literal[
-        "mewh", "mmgeoid", "microGal", "potential", "pascal", "mvcu", "norm"
-    ] = "mewh",
+    unit: str = "mewh",
     errors=False,
     lmax: int | None = None,
     mmax: int | None = None,
@@ -244,7 +242,6 @@ def sh_to_grid(
     data : xr.Dataset
         xr.Dataset that corresponds to SH data to convert into spatial representation.
     unit : str, optional
-        'mewh', 'mmgeoid', 'microGal', 'potential', 'pascal', 'mvcu', or 'norm'
         Unit of the spatial data used in the transformation. Default is 'mewh' for meters of Equivalent Water Height.
         See utils.harmo.l_factor_conv() doc for details on the units.
     errors : bool, optional
@@ -447,9 +444,7 @@ def sh_to_grid(
 def grid_to_sh(
     grid: xr.DataArray,
     lmax: int,
-    unit: Literal[
-        "mewh", "mmgeoid", "microGal", "potential", "pascal", "mvcu", "norm"
-    ] = "mewh",
+    unit: str = "mewh",
     mmax: int | None = None,
     lmin: int = 0,
     mmin: int = 0,
@@ -1227,6 +1222,49 @@ def load_default_love_numbers() -> xr.Dataset:
     return ds
 
 
+def _prefix_parser(unit: str) -> float:
+    """
+    Parse a SI-style prefix from a unit string and return its numeric multiplier.
+
+    Parameters
+    ----------
+    unit : str
+        Unit name that may include a prefix. The function accepts common ASCII and Unicode forms for the micro prefix
+        (both 'u' and 'µ') and is case\-sensitive where relevant (e.g. 'M' for mega vs 'm' for milli).
+
+        Unit cannot be defined using 'm' for milli because it is used for meter unit,
+        so 'milli' or 'mm' should be used instead. Same for 'p' for pico, 'pico' should be used instead.
+
+        Goes from 'giga' / 'G' to 'pico', including 'mega', 'kilo', 'hecto', 'deca', 'deci', 'centi', 'milli',
+        'micro', and 'nano'.
+
+    Returns
+    -------
+    scale : float
+        Multiplicative factor corresponding to the detected prefix.
+        If no recognized prefix is found, returns 1.0 (indicating no scaling).
+    """
+    prefixes = {
+        ("G", "giga"): 1e-9,
+        ("M", "mega"): 1e-6,
+        ("k", "kilo"): 1e-3,
+        ("h", "hecto"): 1e-2,
+        ("deca",): 1e-1,
+        ("d", "deci"): 1e1,
+        ("c", "centi"): 1e2,
+        ("milli", "mm"): 1e3,  # cannot accept 'm' because it is used for meter unit
+        ("µ", "u", "micro"): 1e6,
+        ("n", "nano"): 1e9,
+        ("pico",): 1e12,  # cannot accept 'p' because it is used for pressure unit
+    }
+
+    for prefix, multiplier in prefixes.items():
+        if unit.startswith(prefix):
+            return multiplier
+
+    return 1.0
+
+
 def _compute_l_factor(
     l: np.ndarray | xr.DataArray,
     unit: str,
@@ -1249,7 +1287,7 @@ def _compute_l_factor(
     l: np.ndarray or xr.DataArray
         Degrees
     unit: str
-        Unit type for conversion
+        Unit type for conversion ending with the unit name
     geocentric_colat : xr.DataArray
         Geocentric colatitude
     ds_love: xr.Dataset, optional
@@ -1277,7 +1315,7 @@ def _compute_l_factor(
         Degree-dependent conversion factor.
     """
[…]
