### 2026-02-25 38f1546
Vectorization of _scale_plm_factors() to gain a factor two in time computation of compute_plm()
