+++ lenapy/utils/harmo.py
@@ -225,6 +225,7 @@ def sh_to_grid(
     include_elastic: bool = True,
     plm: xr.DataArray = None,
     normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
+    dtype_plm: type[complex] | type[float] = np.float128,
     use_dask: bool = False,
     chunks_lfactor: dict | None = None,
     chunks_plm: dict | None = None,
@@ -299,6 +300,8 @@ def sh_to_grid(
         If plm need to be computed, choice of the norm corresponding to the SH dataset.
         Either '4pi', 'ortho', or 'schmidt' for 4pi normalized, orthonormalized, or Schmidt semi-normalized SH
         functions, respectively. Default is '4pi'.
+    dtype_plm : dtype, optional
+        Specify the dtype to compute the plm DataArray. Default is np.float128.
 
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
@@ -351,6 +354,7 @@ def sh_to_grid(
             latitude=latitude,
             mmax=mmax,
             normalization=normalization_plm,
+            dtype=dtype_plm,
             use_dask=use_dask,
             chunks=chunks_plm,
         )
@@ -454,6 +458,7 @@ def grid_to_sh(
     include_elastic: bool = True,
     plm: xr.DataArray | None = None,
     normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
+    dtype_plm: type[complex] | type[float] = np.float128,
     use_dask: bool = False,
     chunks_plm: dict | None = None,
     **kwargs,
@@ -500,6 +505,8 @@ def grid_to_sh(
         If plm need to be computed, choice of the norm.  Either '4pi', 'ortho', or 'schmidt' for
         4pi normalized, orthonormalized, or Schmidt semi-normalized SH functions, respectively. Default is '4pi'.
         Output SH coefficient will be normalized according to this parameter.
+    dtype_plm : dtype, optional
+        Specify the dtype to compute the plm DataArray. Default is np.float128.
 
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
@@ -563,6 +570,7 @@ def grid_to_sh(
             latitude=grid.cf["latitude"],
             mmax=mmax,
             normalization=normalization_plm,
+            dtype=dtype_plm,
             use_dask=use_dask,
             chunks=chunks_plm,
         )
@@ -613,10 +621,14 @@ def _scale_plm_factors(
     lmax: int,
     normalization: Literal["4pi", "ortho", "schmidt"],
     derivative: bool,
+    dtype: complex | float | type[complex] | type[float],
 ) -> tuple[np.ndarray, np.ndarray, float, float, Optional[np.ndarray]]:
     """
     Compute recurrence coefficients f1 and f2 for the Legendre recursion.
     Compute df recurrence coefficients for the derivative if requested.
+
+    Operation used here are the developed and vectorize form of the SHTOOLS implementation
+    to minimize the use of sqrt and division operations.
 
     Parameters
     ----------
@@ -626,6 +638,8 @@ def _scale_plm_factors(
         Normalization scheme.
     derivative : bool
         Whether to compute factors for the derivative of plm. Default is False.
+    dtype : dtype
+        Data type of the output arrays.
 
     Returns
     -------
@@ -641,61 +655,71 @@ def _scale_plm_factors(
         Recurrence factor for the derivative of plm, only returned if derivative is True.
     """
     size = (lmax + 1) * (lmax + 2) // 2
[…]
