### 2026-02-27 613f8a3
Vectorization of plm legendre estimation to gain a factor ten in time computation of compute_plm()
