+++ lenapy/utils/gravity.py
@@ -549,9 +549,9 @@ def sh_to_gravity_disturbance(
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
     chunks_lfactor : dict, optional
-        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 20}.
+        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 200}.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 4}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
@@ -1058,9 +1058,9 @@ def sh_to_potential_partial_derivative_longitude(
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
     chunks_lfactor : dict, optional
-        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 20}.
+        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 200}.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 4}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
