+++ lenapy/utils/harmo.py
@@ -306,9 +306,9 @@ def sh_to_grid(
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
     chunks_lfactor : dict, optional
-        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 20}.
+        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 200}.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 4}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
@@ -460,6 +460,7 @@ def grid_to_sh(
     normalization_plm: Literal["4pi", "ortho", "schmidt"] = "4pi",
     dtype_plm: type[complex] | type[float] = np.float128,
     use_dask: bool = False,
+    chunks_lfactor: dict | None = None,
     chunks_plm: dict | None = None,
     **kwargs,
 ) -> xr.Dataset:
@@ -510,8 +511,10 @@ def grid_to_sh(
 
     use_dask : bool, optional
         If True, use dask to chunk plm for memory optimization. Default is False.
+    chunks_lfactor : dict, optional
+        Define the chunking of lfactor when use_dask is True. Default is None, which set the chunking to {'l': 400}.
     chunks_plm : dict, optional
-        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 10}.
+        Define the chunking of plm when use_dask is True. Default is None, which set the chunking to {'latitude': 4}.
 
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
@@ -558,6 +561,8 @@ def grid_to_sh(
         ellipsoidal_earth=ellipsoidal_earth,
         geocentric_colat=geocentric_colat,
         attrs=grid.attrs,
+        use_dask=use_dask,
+        chunks=chunks_lfactor,
         **kwargs,
     )
 
@@ -643,83 +648,67 @@ def _scale_plm_factors(
 
     Returns
     -------
-    f1 : np.ndarray
+    f1 : np.ndarray (lmax+1, lmax+1)
         Recurrence factor f1.
-    f2 : np.ndarray
+    f2 : np.ndarray (lmax+1, lmax+1)
         Recurrence factor f2.
     norm_p10 : float
         Normalization for P(1,0).
     norm_4pi : float
         Overall normalization factor.
-    df : np.ndarray | None
+    df : np.ndarray (lmax+1, lmax+1) | None
         Recurrence factor for the derivative of plm, only returned if derivative is True.
     """
-    size = (lmax + 1) * (lmax + 2) // 2
-    f1 = np.zeros(size, dtype=dtype)
-    f2 = np.zeros(size, dtype=dtype)
-    df = np.zeros(size, dtype=dtype) if derivative else None
+    l = np.tile(np.arange(lmax + 1)[:, None], (1, lmax + 1))
+    m = np.tile(np.arange(lmax + 1)[None, :], (lmax + 1, 1))
+    mask = m < l - 1
+    l_mask = l[mask]
+    m_mask = m[mask]
+    f1 = np.zeros((lmax + 1, lmax + 1), dtype=dtype)
[…]
