### 2026-02-27 fc7dd08
Change compute_plm() test
