### 2026-02-27 83eff15
Change compute_plm() test bis
