### 2026-03-03 b07a657
Refactor code with earth_radius() function One small debug on centrifugal derivative estimation Add height gestion for normal gravity estimation
