+++ lenapy/utils/gravity.py
@@ -34,7 +34,11 @@ import numpy as np
 import xarray as xr
 
 from lenapy.constants import *
-from lenapy.utils.geo import assert_latitude, latitude_to_geocentric_colatitude
+from lenapy.utils.geo import (
+    assert_latitude,
+    earth_radius,
+    latitude_to_geocentric_colatitude,
+)
 from lenapy.utils.harmo import (
     _assert_plm,
     _generate_grid,
@@ -330,7 +334,9 @@ def change_love_reference_frame(
 
 
 def estimate_normal_gravity(
-    geographic_latitude: xr.DataArray | np.ndarray | None = None,
+    height: xr.DataArray | None = None,
+    radius: xr.DataArray | None = None,
+    latitude: xr.DataArray | np.ndarray | None = None,
     a_earth: float | None = LNPY_A_EARTH_GRS80,
     earth_gravity_constant: float | None = LNPY_GM_EARTH,
     f_earth: float = LNPY_F_EARTH_GRS80,
@@ -342,8 +348,17 @@ def estimate_normal_gravity(
 
     Parameters
     ----------
-    geographic_latitude: xr.DataArray | np.ndarray, optional
-        Geographic / geodetic latitude for ellipsoidal Earth radius computation in radians.
+    height: xr.DataArray | float | None, optional
+        Height above of the surface of reference (sphere or ellipsoid) depending on latitude and possibly other
+        dimensions. If not provided, radius parameter is used for the computation.
+        Used in priority over radius parameter if both are provided.
+    radius: xr.DataArray | None, optional
+        Radius from the Earth's center depending on latitude and other dimensions. If not provided, `height` parameter
+        is used to compute the radius.
+    latitude: xr.DataArray | np.ndarray, optional
+        To provide if no height or radius is given to compute the normal gravity at the
+        surface of reference (sphere or ellipsoid).
+        Geographic latitude in degree.
     a_earth : float, optional
         Earth semi-major axis [m]. Default is LNPY_A_EARTH_GRS80.
     earth_gravity_constant : float, optional
@@ -358,6 +373,25 @@ def estimate_normal_gravity(
     gamma_0: xr.DataArray | np.ndarray
         Normal gravity at given geographic colatitude.
     """
+    if height is None and radius is not None:
+        assert_latitude(radius)
+        latitude = radius.latitude
+    elif height is not None:
+        assert_latitude(height)
+        latitude = height.latitude
+    elif latitude is None:
+        raise ValueError(
+            "If you do not provide radius or height parameters, you need to provide latitude to "
+            "compute the normal gravity at these latitude on the ellipsoid of reference."
+        )
+
+    if radius is not None and height is None:
+        height = radius - earth_radius(latitude, True, a_earth, f_earth)
+
+    geoc_colat = latitude_to_geocentric_colatitude(
+        latitude, ellipsoidal_earth=True, f_earth=f_earth
+    )
+
     e_prime = np.sqrt(2 * f_earth - f_earth**2) / (1 - f_earth)
     q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
     q0_prime = 3 * (1 + 1 / e_prime**2) * (1 - 1 / e_prime * np.arctan(e_prime)) - 1
@@ -373,13 +407,21 @@ def estimate_normal_gravity(
         earth_gravity_constant / a_earth**2 * (1 + m * e_prime * q0_prime / 3 / q0)
     )
 
-    return (
-        gamma_e * np.cos(geographic_latitude) ** 2
[…]
