### 2026-03-03 d759487
small debug for height gestion for normal gravity estimation
