### 2026-03-05 fc062f1
Temporal save of tests on normal gravity
