+++ lenapy/utils/gravity.py
@@ -341,6 +341,10 @@ def estimate_normal_gravity(
     earth_gravity_constant: float | None = LNPY_GM_EARTH,
     f_earth: float = LNPY_F_EARTH_GRS80,
     omega_earth: float = LNPY_OMEGA_EARTH_GRS80,
+    mode: Literal[
+        "somigliana", "gravi_disturbance", "gravi_disturbance_centri"
+    ] = "somigliana",
+    lmax=None,
 ) -> xr.DataArray:
     """
     Estimate normal acceleration with the Somigliana equation at given geographic latitude
@@ -390,38 +394,155 @@ def estimate_normal_gravity(
     elif height is None:
         height = 0
 
-    geoc_colat = latitude_to_geocentric_colatitude(
+    geographic_colat = latitude_to_geocentric_colatitude(
+        latitude, ellipsoidal_earth=False, f_earth=f_earth
+    )
+
+    geocentric_colat = latitude_to_geocentric_colatitude(
         latitude, ellipsoidal_earth=True, f_earth=f_earth
     )
 
-    e_prime = np.sqrt(2 * f_earth - f_earth**2) / (1 - f_earth)
-    q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
-    q0_prime = 3 * (1 + 1 / e_prime**2) * (1 - 1 / e_prime * np.arctan(e_prime)) - 1
-    m = omega_earth**2 * a_earth**3 * (1 - f_earth) / earth_gravity_constant
+    if "somigliana" in mode:
+        e_prime = np.sqrt(2 * f_earth - f_earth**2) / (1 - f_earth)
+        q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
+        q0_prime = 3 * (1 + 1 / e_prime**2) * (1 - 1 / e_prime * np.arctan(e_prime)) - 1
+        m = omega_earth**2 * a_earth**3 * (1 - f_earth) / earth_gravity_constant
 
-    gamma_e = (
-        earth_gravity_constant
-        / a_earth**2
-        / (1 - f_earth)
-        * (1 - m - m * e_prime * q0_prime / 6 / q0)
-    )
-    gamma_p = (
-        earth_gravity_constant / a_earth**2 * (1 + m * e_prime * q0_prime / 3 / q0)
-    )
+        gamma_e = (
+            earth_gravity_constant
+            / (a_earth**2 * (1 - f_earth))
+            * (1 - m - m * e_prime * q0_prime / (6 * q0))
+        )
+        gamma_p = (
+            earth_gravity_constant
+            / a_earth**2
+            * (1 + m * e_prime * q0_prime / (3 * q0))
+        )
 
-    gamma_0 = (
-        gamma_e * np.cos(geoc_colat) ** 2
-        + (1 - f_earth) * gamma_p * np.sin(geoc_colat) ** 2
-    ) / np.sqrt(np.cos(geoc_colat) ** 2 + (1 - f_earth) ** 2 * np.sin(geoc_colat) ** 2)
+        if "reduced" in mode:
+            gamma_0 = (
+                gamma_e
+                * (
+                    1
+                    + ((1 - f_earth) * gamma_p / gamma_e - 1)
+                    * np.cos(geographic_colat) ** 2
+                )
+                / np.sqrt(
+                    1 - (2 * f_earth - f_earth**2) * np.cos(geographic_colat) ** 2
+                )
+            )
 
-    gamma = gamma_0 * (
-        1
-        - 2
-        * (1 + f_earth + m - 2 * f_earth * np.sin(geoc_colat) ** 2)
-        * height
-        / a_earth
[…]
