+++ lenapy/utils/gravity.py
@@ -398,151 +398,37 @@ def estimate_normal_gravity(
         latitude, ellipsoidal_earth=False, f_earth=f_earth
     )
 
-    geocentric_colat = latitude_to_geocentric_colatitude(
-        latitude, ellipsoidal_earth=True, f_earth=f_earth
+    e_prime = np.sqrt(2 * f_earth - f_earth**2) / (1 - f_earth)
+    q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
+    q0_prime = 3 * (1 + 1 / e_prime**2) * (1 - 1 / e_prime * np.arctan(e_prime)) - 1
+    m = omega_earth**2 * a_earth**3 * (1 - f_earth) / earth_gravity_constant
+
+    gamma_e = (
+        earth_gravity_constant
+        / (a_earth**2 * (1 - f_earth))
+        * (1 - m - m * e_prime * q0_prime / (6 * q0))
+    )
+    gamma_p = (
+        earth_gravity_constant / a_earth**2 * (1 + m * e_prime * q0_prime / (3 * q0))
     )
 
-    if "somigliana" in mode:
-        e_prime = np.sqrt(2 * f_earth - f_earth**2) / (1 - f_earth)
-        q0 = (0.5 + 1.5 / e_prime**2) * np.arctan(e_prime) - 1.5 / e_prime
-        q0_prime = 3 * (1 + 1 / e_prime**2) * (1 - 1 / e_prime * np.arctan(e_prime)) - 1
-        m = omega_earth**2 * a_earth**3 * (1 - f_earth) / earth_gravity_constant
+    gamma_0 = (
+        gamma_e * np.sin(geographic_colat) ** 2
+        + (1 - f_earth) * gamma_p * np.cos(geographic_colat) ** 2
+    ) / np.sqrt(
+        np.sin(geographic_colat) ** 2
+        + (1 - f_earth) ** 2 * np.cos(geographic_colat) ** 2
+    )
 
-        gamma_e = (
-            earth_gravity_constant
-            / (a_earth**2 * (1 - f_earth))
-            * (1 - m - m * e_prime * q0_prime / (6 * q0))
-        )
-        gamma_p = (
-            earth_gravity_constant
-            / a_earth**2
-            * (1 + m * e_prime * q0_prime / (3 * q0))
-        )
-
-        if "reduced" in mode:
-            gamma_0 = (
-                gamma_e
-                * (
-                    1
-                    + ((1 - f_earth) * gamma_p / gamma_e - 1)
-                    * np.cos(geographic_colat) ** 2
-                )
-                / np.sqrt(
-                    1 - (2 * f_earth - f_earth**2) * np.cos(geographic_colat) ** 2
-                )
-            )
-
-        elif "_serie" in mode:
-            beta = 5.3024e-3
-            beta1 = -5.8e-6
-            gamma_0 = gamma_e * (
-                1
-                + beta * np.cos(geographic_colat) ** 2
-                + beta1 * np.cos(2 * geographic_colat) ** 2
-            )
-
-        elif "_GRS80serie" in mode:
-            c1 = 5.2790414e-3
-            c2 = 2.32718e-5
-            c3 = 1.262e-7
-            c4 = 7e-10
-            gamma_0 = gamma_e * (
-                1
-                + c1 * np.cos(geographic_colat) ** 2
-                + c2 * np.cos(geographic_colat) ** 4
-                + c3 * np.cos(geographic_colat) ** 6
-                + c4 * np.cos(geographic_colat) ** 8
-            )
-        else:
-            gamma_0 = (
-                gamma_e * np.sin(geographic_colat) ** 2
[…]
