### 2026-03-07 2d11cce
Refactor gm_earth to earth_gravit_constant
