+++ lenapy/utils/harmo.py
@@ -313,7 +313,7 @@ def sh_to_grid(
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
         for the unit conversion. These parameters include (see :func:`l_factor_conv` documentation for more details) :
-        a_earth, gm_earth, f_earth, omega_earth, rho_earth, ds_love
+        a_earth, earth_gravity_constant, f_earth, omega_earth, rho_earth, ds_love
 
     Returns
     -------
@@ -519,7 +519,7 @@ def grid_to_sh(
     **kwargs :
         Supplementary parameters used by the function l_factor_conv to modify defaults constants used in the computation
         for the unit conversion. These parameters include (see :func:`l_factor_conv` documentation for more details) :
-        a_earth, gm_earth, f_earth, omega_earth, rho_earth, ds_love
+        a_earth, earth_gravity_constant, f_earth, omega_earth, rho_earth, ds_love
 
     Returns
     -------
@@ -1181,7 +1181,7 @@ def change_normalization(
 
 
 def _get_earth_parameters(
-    attrs: dict | None, a_earth: float | None, gm_earth: float | None
+    attrs: dict | None, a_earth: float | None, earth_gravity_constant: float | None
 ) -> tuple[float, float]:
     """
     Retrieve Earth parameters from inputs or fallback constants.
@@ -1192,14 +1192,14 @@ def _get_earth_parameters(
         Attributes potentially containing Earth parameters.
     a_earth : float or None
         Semi-major axis.
-    gm_earth : float or None
+    earth_gravity_constant : float or None
         Gravitational constant.
 
     Returns
     -------
     a_earth : float
         Resolved semi-major axis.
-    gm_earth : float
+    earth_gravity_constant : float
         Resolved gravitational constant.
     """
     if attrs is None:
@@ -1207,7 +1207,9 @@ def _get_earth_parameters(
 
     # return first element if not None, otherwise return value for the attrs.key if it exists otherwise return last term
     resolved_a = a_earth or float(attrs.get("radius", LNPY_A_EARTH_GRS80))
-    resolved_gm = gm_earth or float(attrs.get("earth_gravity_constant", LNPY_GM_EARTH))
+    resolved_gm = earth_gravity_constant or float(
+        attrs.get("earth_gravity_constant", LNPY_GM_EARTH)
+    )
     return resolved_a, resolved_gm
 
 
@@ -1278,7 +1280,7 @@ def _compute_l_factor(
     geocentric_colat: xr.DataArray | None,
     ds_love: xr.Dataset | None,
     a_earth: float,
-    gm_earth: float,
+    earth_gravity_constant: float,
     f_earth: float,
     omega_earth: float,
     rho_earth: float,
@@ -1301,7 +1303,7 @@ def _compute_l_factor(
         Love numbers
     a_earth: float
         Earth radius
-    gm_earth: float
+    earth_gravity_constant: float
         Earth gravitational constant
     f_earth : float
         Earth flattening
@@ -1346,11 +1348,11 @@ def _compute_l_factor(
             gamma_0 = estimate_normal_gravity(
                 latitude=geocentric_colat.latitude,
                 a_earth=a_earth,
-                earth_gravity_constant=gm_earth,
+                earth_gravity_constant=earth_gravity_constant,
                 f_earth=f_earth,
                 omega_earth=omega_earth,
             )
[…]
