### 2026-03-23 3442910
Gestion of dimension with no latitude and longitude, only in coords
