### 2026-03-23 6cbbbca
Rebuild normal_zonal_correction for inside convertion of coefficients depending on output reference surface
Correction of sign + partial for dlatitude derivation

Add deflection of vertical

Add all gravity function to lnharmo
