+++ lenapy/utils/gravity.py
@@ -434,29 +434,41 @@ def estimate_normal_gravity(
 
 def normal_zonal_correction(
     ds: xr.Dataset,
-    a_earth: float | None = None,
-    earth_gravity_constant: float | None = None,
+    a_earth: float | None = LNPY_A_EARTH_GRS80,
+    earth_gravity_constant: float | None = LNPY_GM_EARTH_GRS80,
     f_earth: float = LNPY_F_EARTH_GRS80,
     omega_earth: float = LNPY_OMEGA_EARTH_GRS80,
+    final_a_earth: float | None = None,
+    final_earth_gravity_constant: float | None = None,
     reverse: bool = False,
     apply: bool = True,
 ) -> xr.Dataset:
     """
-    Normal gravity field correction on zonal coefficients on a SH dataset for a specified ellipsoid.
+    Normal gravity field correction on zonal coefficients on a SH dataset for a specified ellipsoid define by a_earth,
+    earth_gravity_constant, f_earth and omega_earth.
+    Default used ellipsoid is GRS80.
+
+    The zonal field coefficients can be converted to a new ellipsoid by giving parameters final_a_earth
+    and final_earth_gravity_constant.
 
     Parameters
     ----------
     ds : xr.Dataset
         xr.Dataset that corresponds to SH data to be correct for the normal gravity field.
     a_earth : float | None, optional
-        Earth radius constant of the dataset ds in meters. If not provided, uses `ds.attrs['radius']`.
+        Earth radius constant of the dataset ds in meters. Default is LNPY_A_EARTH_GRS80.
     earth_gravity_constant : float | None, optional
-        Current gravitational constant of the Earth of the dataset ds in m³/s².
-        If not provided, uses `ds.attrs['earth_gravity_constant']`.
+        Current gravitational constant of the Earth of the dataset ds in m³/s². Default is LNPY_GM_EARTH_GRS80.
     f_earth : float, optional
         Earth flattening. Default is LNPY_F_EARTH_GRS80.
     omega_earth : float, optional
         Earth rotation rate. Default is LNPY_OMEGA_EARTH_GRS80.
+    final_a_earth : float | None, optional
+        Final reference surface radius for normal Stokes coefficients.
+        If not provided, uses `ds.attrs['radius']` and if it does not exist, use a_earth.
+    final_earth_gravity_constant : float | None, optional
+        Final reference gravity constant for normal Stokes coefficients.
+        If not provided, uses `ds.attrs['earth_gravity_constant']` and if it does not exist, use earth_gravity_constant.
     reverse : bool, optional
         False to apply the correction, True to remove the correction.
     apply : bool, optional
@@ -473,21 +485,6 @@ def normal_zonal_correction(
         If the current reference frame constants are not provided and not found in the dataset attributes.
 
     """
-    try:
-        a_earth = ds.attrs["radius"] if a_earth is None else a_earth
-        earth_gravity_constant = (
-            ds.attrs["earth_gravity_constant"]
-            if earth_gravity_constant is None
-            else earth_gravity_constant
-        )
-
-    except KeyError:
-        raise KeyError(
-            "If you provide no information about the current reference constants of your ds dataset using "
-            "'radius' and 'earth_gravity_constant' parameters, those information need to be "
-            "contained in ds.attrs dict as ds.attrs['radius'] and ds.attrs['earth_gravity_constant']."
-        )
-
     if apply:
         ds_out = ds
     else:
@@ -518,6 +515,34 @@ def normal_zonal_correction(
 
[…]
