### 2026-03-23 454ed2c
Add automatic gestion of the case where you want to estimate spatial grid with a certain radius but coeffs radius do not correspond so internal change_reference() is applied
