### 2026-03-23 a52a987
Add gravity grid test
